added error handling for thread

+++ business/threads/usecase.go
@@ -3,6 +3,7 @@ package threads
 import (
 	"context"
 	"disspace/helpers/messages"
+	"strings"
 	"time"
 )
 
@@ -23,7 +24,7 @@ func (useCase *ThreadUseCase) GetAll(ctx context.Context, sort string) ([]Domain
 		return []Domain{}, messages.ErrInvalidQueryParam
 	}
 
-	result, err := useCase.threadRepo.GetAll(ctx, sort) 
+	result, err := useCase.threadRepo.GetAll(ctx, sort)
 	if err != nil {
 		return []Domain{}, err
 	}
@@ -31,6 +32,10 @@ func (useCase *ThreadUseCase) GetAll(ctx context.Context, sort string) ([]Domain
 }
 
 func (useCase *ThreadUseCase) Create(ctx context.Context, threadDomain *Domain) (Domain, error) {
+	if threadDomain.Title == "" || strings.TrimSpace(threadDomain.Title) == "" {
+		return Domain{}, messages.ErrEmptyTitle
+	}
+
 	result, err := useCase.threadRepo.Create(ctx, threadDomain)
 	if err != nil {
 		return Domain{}, err
@@ -39,6 +44,10 @@ func (useCase *ThreadUseCase) Create(ctx context.Context, threadDomain *Domain) 
 }
 
 func (useCase *ThreadUseCase) Delete(ctx context.Context, id string) error {
+	if strings.TrimSpace(id) == "" {
+		return messages.ErrInvalidThreadID
+	}
+
 	err := useCase.threadRepo.Delete(ctx, id)
 	if err != nil {
 		return messages.ErrInvalidThreadID
@@ -47,6 +56,9 @@ func (useCase *ThreadUseCase) Delete(ctx context.Context, id string) error {
 }
 
 func (useCase *ThreadUseCase) GetByID(ctx context.Context, id string) (Domain, error) {
+	if strings.TrimSpace(id) == "" {
+		return Domain{}, messages.ErrInvalidThreadID
+	}
 	result, err := useCase.threadRepo.GetByID(ctx, id)
 	if err != nil {
 		return Domain{}, messages.ErrInvalidThreadID
@@ -55,6 +67,9 @@ func (useCase *ThreadUseCase) GetByID(ctx context.Context, id string) (Domain, e
 }
 
 func (useCase *ThreadUseCase) Update(ctx context.Context, threadDomain *Domain, id string) error {
+	if strings.TrimSpace(id) == "" {
+		return messages.ErrInvalidThreadID
+	}
 	err := useCase.threadRepo.Update(ctx, threadDomain, id)
 	if err != nil {
 		return messages.ErrInvalidThreadID
